requirements.txt
requests
yfinance
python-dotenv
pandas